2-strlen.c file modified

+++ 0x05-pointers_arrays_strings/2-strlen.c
@@ -7,7 +7,7 @@
  */
 int _strlen(char *s)
 {
-	size_t length = 0;
+	int length = 0;
 
 	while (*str++)
 		length++;
